Updates package name and license
