+++ src/main/scala/com/fortysevendeg/katagenda/AgendaApplication.scala
@@ -1,23 +1,24 @@
 /*
- * Copyright (C) 2015 Karumi.
+ * Copyright (C) 2015 47 Degrees, LLC http://47deg.com hello@47deg.com
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
-package com.karumi.katagenda
+package com.fortysevendeg.katagenda
 
-import com.karumi.katagenda.common.ui.{View, Presenter}
-import com.karumi.katagenda.servicelocator.AgendaServiceLocator
-import com.karumi.katagenda.ui.ContactsListPresenter
+import com.fortysevendeg.katagenda.common.ui.{View, Presenter}
+import com.fortysevendeg.katagenda.servicelocator.AgendaServiceLocator
+import com.fortysevendeg.katagenda.ui.ContactsListPresenter
 
 object AgendaApplication extends App {
