+++ src/test/scala/com/fortysevendeg/katagenda/domain/AgendaSpec.scala
@@ -1,7 +1,24 @@
-package com.karumi.katagenda.domain
+/*
+ * Copyright (C) 2015 47 Degrees, LLC http://47deg.com hello@47deg.com
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
 
-import com.karumi.katagenda.common.repository.InMemoryDataSource
-import com.karumi.katagenda.domain.repository.ContactsRepository
+package com.fortysevendeg.katagenda.domain
+
+import com.fortysevendeg.katagenda.common.repository.InMemoryDataSource
+import com.fortysevendeg.katagenda.domain.repository.ContactsRepository
 import org.specs2.mutable.Specification
 import org.specs2.specification.Scope
